refactor: move Player class to its own module
- Extracted Player class from sprites.py into a dedicated player.py module
- Updated all import statements to reference the new location
- This improves code organization and follows the single-responsibility principle
